Fix AkRoom not detecting moving between rooms
1) We need to use Object::cast_to instead of static_cast in GDExtension
2) Object::get_instance_id is a uint64_t, so the room id in the AkEvent3D and AkListener3D needs to be the same

+++ addons/Wwise/native/src/scene/ak_event.cpp
@@ -381,6 +381,6 @@ void AkEvent3D::set_is_environment_aware(bool is_environment_aware)
 
 bool AkEvent3D::get_is_environment_aware() const { return is_environment_aware; }
 
-void AkEvent3D::set_room_id(int room_id) { this->room_id = room_id; }
+void AkEvent3D::set_room_id(uint64_t room_id) { this->room_id = room_id; }
 
-int AkEvent3D::get_room_id() const { return room_id; }
+uint64_t AkEvent3D::get_room_id() const { return room_id; }
